feat: Refactor interfaces into own package
