--- packages/backend/src/interfaces/game/UserIdRoomMap.ts
@@ -1,8 +1,0 @@
-interface UserIdRoomMapData {
-  room: string;
-  username: string;
-}
-
-export default interface UserIdRoomMap {
-  [key: string]: UserIdRoomMapData;
-}
