+++ packages/frontend/src/contexts/GameContext.tsx
@@ -1,6 +1,6 @@
 import { createContext } from "react";
 
-import Room from "interfaces/Room";
+import { Room } from "werewolf-types";
 
 const GameContext = createContext({
   username: "",
