+++ packages/frontend/src/contexts/sockets/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import io from "socket.io-client";
 import { socketEvents } from "./listen";
-import Room from "interfaces/Room";
+import { Room } from "werewolf-types";
 
 export const socket = io("http://localhost:8666");
