--- packages/frontend/src/interfaces/Room.ts
@@ -1,3 +1,0 @@
-export default interface Room {
-  connectedUsers: Array<string>
-};
