--- packages/frontend/src/interfaces/socket/EnterRoom.ts
@@ -1,11 +1,0 @@
-export interface EnterRoomResponse {
-  userId: string;
-  connectedUsers: Array<string>;
-}
-
-export interface EnterRoomPayload {
-  room: string;
-  userId: string;
-  username: string;
-  message: string;
-}
