--- packages/frontend/src/interfaces/socket/ErrorResponse.ts
@@ -1,4 +1,0 @@
-export default interface ErrorResponse {
-  errorCode: number;
-  errorMessage: string;
-}
