+++ packages/types/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "werewolf-types",
+  "version": "0.1.0",
+  "private": true,
+  "license": "MIT"
+}
